Edited delete logic and design

+++ lib/shared/components/components.dart
@@ -39,19 +39,73 @@ Widget defaultTextFormField({
 
 Widget taskItem(Map model, context) => Dismissible(
       key: Key(model['id'].toString()),
+      background: Row(
+        children: [
+          Expanded(
+            child: Container(
+              padding: const EdgeInsets.all(20.0),
+              margin: const EdgeInsets.only(
+                top: 10.0,
+                bottom: 10.0,
+              ),
+              alignment: AlignmentDirectional.centerStart,
+              decoration: const BoxDecoration(
+                color: Colors.red,
+                borderRadius: BorderRadius.only(
+                  topLeft: Radius.circular(
+                    10.0,
+                  ),
+                  bottomLeft: Radius.circular(
+                    10.0,
+                  ),
+                ),
+              ),
+              child: const Icon(
+                Icons.delete,
+                color: Colors.white,
+              ),
+            ),
+          ),
+          Expanded(
+            child: Container(
+              padding: const EdgeInsets.all(20.0),
+              margin: const EdgeInsets.only(
+                top: 10.0,
+                bottom: 10.0,
+              ),
+              alignment: AlignmentDirectional.centerEnd,
+              decoration: const BoxDecoration(
+                color: Colors.red,
+                borderRadius: BorderRadius.only(
+                  topRight: Radius.circular(
+                    10.0,
+                  ),
+                  bottomRight: Radius.circular(
+                    10.0,
+                  ),
+                ),
+              ),
+              child: const Icon(
+                Icons.delete,
+                color: Colors.white,
+              ),
+            ),
+          ),
+        ],
+      ),
       child: Container(
         padding: const EdgeInsets.all(20.0),
         margin: const EdgeInsets.only(
           top: 10.0,
           bottom: 10.0,
-          right: 20.0,
-          left: 20.0,
+          // right: 20.0,
+          // left: 20.0,
         ),
         decoration: const BoxDecoration(
           gradient: LinearGradient(
             begin: Alignment.topLeft,
             end: Alignment.bottomRight,
-            colors: [Color(0xFF2196F3), Color(0x00252e41)],
+            colors: [Color(0xFF2196F3), Colors.transparent],
           ),
           borderRadius: BorderRadius.all(
             Radius.circular(
@@ -94,9 +148,43 @@ Widget taskItem(Map model, context) => Dismissible(
           ],
         ),
       ),
-      onDismissed: (direction) {
-        AppCubit.get(context).deleteDB(
-          id: model['id'],
+      onDismissed: (direction) {},
+      confirmDismiss: (direction) {
+        return showDialog(
+          context: context,
+          builder: (newContext) {
+            return AlertDialog(
+              shape: const RoundedRectangleBorder(
+                borderRadius: BorderRadius.all(
+                  Radius.circular(
+                    20.0,
+                  ),
+                ),
+              ),
+              title: const Text('Delete Task?'),
+              content: const Text('Task will be deleted permanently'),
+              actions: [
+                TextButton(
+                  onPressed: () {
+                    AppCubit.get(context).deleteDB(
+                      id: model['id'],
+                    );
+                    return Navigator.pop(context);
+                  },
+                  child: const Text(
+                    'DELETE',
+                    style: TextStyle(
+                      color: Colors.red,
+                    ),
+                  ),
+                ),
+                TextButton(
+                  onPressed: () => Navigator.of(context).pop(false),
+                  child: const Text('CANCEL'),
+                ),
+              ],
+            );
+          },
         );
       },
     );
@@ -113,8 +201,7 @@ Widget addTaskControls(Map model, context) {
       icon: doneIcon(),
       tooltip: 'Mark as done',
     );
-  }
-  else if (model['status'] == 'done') {
+  } else if (model['status'] == 'done') {
     return IconButton(
       onPressed: () {
         AppCubit.get(context).updateDB(
@@ -154,22 +241,31 @@ Widget addTaskControls(Map model, context) {
 }
 
 Icon doneIcon() => const Icon(
-  Icons.check_circle,
-  color: Color(0xFF0078eb),
-);
+      Icons.check_circle,
+      color: Color(0xFF0078eb),
+    );
+
 Icon archiveIcon() => const Icon(
-  Icons.archive,
-  color: Color(0xFF0078eb),
-);
+      Icons.archive,
+      color: Color(0xFF0078eb),
+    );
 
 Widget tasksBuilder({
   required List<Map> tasks,
   required AppStates state,
 }) =>
     tasks.isNotEmpty
-        ? ListView.builder(
-            itemBuilder: (context, index) => taskItem(tasks[index], context),
-            itemCount: tasks.length,
+        ? Container(
+            padding: const EdgeInsets.only(
+              right: 10.0,
+              left: 10.0,
+            ),
+            color: Colors.transparent,
+            child: ListView.builder(
+              itemBuilder: (context, index) => taskItem(tasks[index], context),
+              itemCount: tasks.length,
+              physics: BouncingScrollPhysics(),
+            ),
           )
         : (state is AppGetDatabaseState ||
                 state is AppChangeBottomNavBarState ||
